Title: A/B 출력, Time: 119ms, Memory: 10MB, Status: Wrong Answer - Codetree

+++ 250422/A/B 출력/a-divide-b.java
@@ -8,7 +8,7 @@ public class Main {
         int a = sc.nextInt();
         int b = sc.nextInt();
 
-        double result = a / b;
+        double result = (double) a / b;
 
         System.out.printf("%.21f", result);
     }
